Minimize product_amount_1 for recurring payment plan button
The value in product_amount_1 is charged immediately no matter what. For
one-time payments this is what we want. For recurring payment plans, we
would like to charge nothing to sign up, and only charge the monthly fee
as defined in the plan definition. However, the simplicity of the
QuickLink API requires that we charge something to validate the credit
card in order to sign up for the recurring plan. We've validated that
$0.01 is a valid amount for this purpose.